+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [0.1.2](https://github.com/dbalsom/martypc/releases/tag/0.1.2) (2023-06-XX)
 
 * Redesigned CGA card with 'dynamic clocking' support. Card will now switch between clocking by cycle or character as appropriate.
+* Added menu options to save changes to loaded floppy image(s).
 * Fixed CPU cycle tracelogging
 * Added port mirrors for CGA (thanks th3bar0n)
 * Fixed address wrapping for graphics modes (thanks th3bar0n)
@@ -17,7 +18,6 @@
 * Fixed bug in PIT latch logic (thanks 640KB)
 * Added CTRL-ALT-DEL menu option
 * Known issues
-    * Floppy images are read-only.
     * Hitting a key during boot on a 5160 machine can halt the CPU.
 
 ## [0.1.1](https://github.com/dbalsom/martypc/releases/tag/0.1.1) (2023-05-31)